[217] Fixing review comments
