+++ gimel-dataapi/gimel-common/src/main/scala/com/paypal/gimel/common/conf/GimelConstants.scala
@@ -233,7 +233,8 @@ object GimelConstants {
   val CONNECTION_TIMEOUT = 180
 
   // GTS
-  val GTS_DEFAULT_USER = ""
+  val GTS_DEFAULT_USER_FLAG = "spark.gimel.gts.default.user"
+  val GTS_DEFAULT_USER = "livy"
   val GTS_USER_CONFIG = "gimel.gts.user"
   val GTS_DEFAULT_USER_FLAG = "spark.gimel.gts.default.user"
   val GTS_IMPERSONATION_FLAG = "spark.gimel.gts.impersonation.enabled"
