+++ gimel-dataapi/gimel-common/src/main/scala/com/paypal/gimel/common/utilities/DataSetUtils.scala
@@ -235,6 +235,87 @@ object DataSetUtils {
   }
 
   /**
+    * Fetch the Type of DataSetType based on the DataSetProperties that is Supplied
+    *
+    * @param dataSetProperties DataSetProperties
+    * @return DataSetType
+    */
+  def getSystemType(dataSetProperties: DataSetProperties): (DataSetType.Value) = {
+    dataSetProperties.datasetType.toUpperCase() match {
+      case "HBASE" =>
+        DataSetType.HBASE
+      case "KAFKA" =>
+        val kafkaApiVersion = GenericUtils.getValue(dataSetProperties.props,
+          GimelConstants.GIMEL_KAFKA_VERSION, defaultValue = GimelConstants.GIMEL_KAFKA_DEFAULT_VERSION)
+        if (kafkaApiVersion.equals(GimelConstants.GIMEL_KAFKA_VERSION_ONE)) {
+          DataSetType.KAFKA
+        } else {
+          DataSetType.KAFKA2
+        }
+      case "ELASTIC_SEARCH" =>
+        DataSetType.ES
+      case "JDBC" =>
+        DataSetType.JDBC
+      case "CASSANDRA" =>
+        DataSetType.CASSANDRA
+      case "AEROSPIKE" =>
+        DataSetType.AEROSPIKE
+      case "DRUID" =>
+        DataSetType.DRUID
+      case "HDFS" =>
+        DataSetType.HDFS
+      case "RESTAPI" =>
+        DataSetType.RESTAPI
+      case "HIVE" =>
+        DataSetType.HIVE
+      case "SFTP" =>
+        DataSetType.SFTP
+      case _ =>
+        DataSetType.HIVE
+    }
+  }
+
+  /**
+    * Checks whether the dataSet is HIVE by scanning the pcatalog phrase and also expecting to have the db and table
+    * names to decide it is a HIVE table
+    *
+    * @param dataSet DataSet
+    * @return Boolean
+    */
+  def isStorageTypeUnknown
+  (dataSet: String): Boolean = {
+    dataSet.split('.').head.toLowerCase() != GimelConstants.PCATALOG_STRING && dataSet.split('.').length == 2
+  }
+
+  /**
+    * For a given dataset (table) this function calls getDataSetProperties which calls catalog provider and returns the dataset properties
+    * which will be used to identify the storage of the dataset
+    *
+    * @param datasetName  - Incoming dataset
+    * @param sparkSession - spark session
+    * @param options      - Set of user options
+    * @return - Returns the storage system (hive/teradata/kafka...)
+    */
+  def getSystemType(datasetName: String, sparkSession: SparkSession,
+                    options: Map[String, String]): com.paypal.gimel.common.utilities.DataSetType.Value = {
+    logger.info("Data set name is  ==> " + datasetName)
+    val formattedProps: Map[String, Any] = getProps(options) ++
+      Map(CatalogProviderConfigs.CATALOG_PROVIDER ->
+        sparkSession.conf.get(CatalogProviderConfigs.CATALOG_PROVIDER,
+          CatalogProviderConstants.PRIMARY_CATALOG_PROVIDER))
+
+    // if storage type unknown we will default to HIVE PROVIDER
+    if (isStorageTypeUnknown(datasetName)) {
+      formattedProps ++ Map(CatalogProviderConfigs.CATALOG_PROVIDER -> CatalogProviderConstants.HIVE_PROVIDER)
+    }
+
+    val dataSetProperties: DataSetProperties = CatalogProvider.getDataSetProperties(datasetName, options)
+    logger.info("dataSetProperties  ==> " + dataSetProperties.toString())
+    val systemType = getSystemType(dataSetProperties)
+    systemType
+  }
+
+  /**
    * Sets log level based on the input from user
    *
    * @param sparkSession
