+++ gimel-dataapi/gimel-common/src/main/scala/com/paypal/gimel/datasetfactory/GimelDataSet.scala
@@ -143,7 +143,7 @@ abstract class GimelDataSet(sparkSession: SparkSession) {
    *                props is the way to set various additional parameters for read and write operations in DataSet class
    *                Example Usecase : to read kafka from-to a certain offset range : One can set something like below -
    *                val props = Map("fromOffset" -> 10, "toOffset" -> 20)
-   *                val data = Dataset(sc).read("wuser.topic", props)
+   *                val data = Dataset(sc).read("flights.topic", props)
    * @return Unit => Empty implementation
    */
   def authenticate(dataset: String, dataSetProps: Map[String, Any] = Map.empty): Unit = {
