+++ gimel-dataapi/gimel-connectors/gimel-elasticsearch-6.2/src/main/scala/com/paypal/gimel/elasticsearch/DataSet.scala
@@ -130,7 +130,6 @@ class DataSet(sparkSession: SparkSession) extends GimelDataSet(sparkSession: Spa
         val errorMessage =
           s"""
              |Writing to dynamic dataset with partitioned index is not currently supported
-             |Solutions for common exceptions are documented here : http://go/gimel/exceptions"
              |""".stripMargin
         throw new Exception(errorMessage)
