+++ gimel-dataapi/gimel-core/src/main/scala/com/paypal/gimel/DataSet.scala
@@ -205,7 +205,7 @@ class DataSet(val sparkSession: SparkSession) {
         // throw error to console
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(e.toString + "\n" + msg, e)
     }
   }
@@ -339,7 +339,7 @@ class DataSet(val sparkSession: SparkSession) {
         // throw error to console
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(e.toString + "\n" + msg, e)
     }
 
@@ -455,7 +455,7 @@ class DataSet(val sparkSession: SparkSession) {
         // throw error to console
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(msg, e)
     }
 
@@ -642,7 +642,7 @@ class DataSet(val sparkSession: SparkSession) {
         // throw error to console
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(e.getMessage + "\n" + msg, e)
     }
   }
@@ -772,7 +772,7 @@ class DataSet(val sparkSession: SparkSession) {
 
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(e.getMessage + "\n" + msg, e)
     }
   }
@@ -901,7 +901,7 @@ class DataSet(val sparkSession: SparkSession) {
         // throw error to console
         logger.throwError(e.toString)
 
-        val msg = s"Error in DataSet ${MethodName} Operation. Common Gimel 'Exceptions' are explained here : http://go/gimel/exceptions"
+        val msg = s"Error in DataSet ${MethodName} Operation."
         throw new DataSetOperationException(e.getMessage + "\n" + msg, e)
     }
 
@@ -969,72 +969,6 @@ object DataSetUtils {
   }
 
   /**
-   * Fetch the Type of DataSetType based on the DataSetProperties that is Supplied
-   *
-   * @param dataSetProperties DataSetProperties
-   * @return DataSetType
-   */
-
-  def getSystemType(dataSetProperties: DataSetProperties): (DataSetType.Value) = {
-
-    val storageHandler = dataSetProperties.props.getOrElse(GimelConstants.STORAGE_HANDLER, GimelConstants.NONE_STRING)
-    val storageType = dataSetProperties.datasetType
-    val kafkaApiVersion = GenericUtils.getValue(dataSetProperties.props,
-      GimelConstants.GIMEL_KAFKA_VERSION, defaultValue = GimelConstants.GIMEL_KAFKA_DEFAULT_VERSION)
-
-    val systemType = storageHandler match {
-      case HbaseConfigs.hbaseStorageHandler =>
-        DataSetType.HBASE
-      case ElasticSearchConfigs.esStorageHandler =>
-        DataSetType.ES
-      case KafkaConfigs.kafkaStorageHandler =>
-        DataSetType.KAFKA2
-      case JdbcConfigs.jdbcStorageHandler =>
-        DataSetType.JDBC
-      case _ =>
-        storageType.toUpperCase() match {
-          case "HBASE" =>
-            DataSetType.HBASE
-          case "SHERLOCK" =>
-            DataSetType.SHERLOCK
-          case "KAFKA" =>
-            if (kafkaApiVersion.equals(GimelConstants.GIMEL_KAFKA_VERSION_ONE)) {
-              DataSetType.KAFKA
-            } else {
-              DataSetType.KAFKA2
-            }
-          case "ELASTIC_SEARCH" =>
-            DataSetType.ES
-          case "JDBC" =>
-            DataSetType.JDBC
-          case "CASSANDRA" =>
-            DataSetType.CASSANDRA
-          case "AEROSPIKE" =>
-            DataSetType.AEROSPIKE
-          case "DRUID" =>
-            DataSetType.DRUID
-          case "HDFS" =>
-            DataSetType.HDFS
-          case "RESTAPI" =>
-            DataSetType.RESTAPI
-          case "PIT" =>
-            DataSetType.PIT
-          case "MONGODB" =>
-            DataSetType.MONGODB
-          case "HIVE" =>
-            DataSetType.HIVE
-          case "SFTP" =>
-            DataSetType.SFTP
-          case "S3" =>
-            DataSetType.S3
-          case _ =>
-            DataSetType.HIVE
-        }
-    }
-    systemType
-  }
-
-  /**
     * provides an appropriate gimel DataSet
     *
     * @param sparkSession : SparkSession
@@ -1091,45 +1025,6 @@ object DataSetUtils {
     }.toOption
   }
 
-  /**
-   * Checks whether the dataSet is HIVE by scanning the pcatalog phrase and also expecting to have the db and table
-   * names to decide it is a HIVE table
-   *
-   * @param dataSet DataSet
-   * @return Boolean
-   */
-  def isStorageTypeUnknown
-  (dataSet: String): Boolean = {
-    dataSet.split('.').head.toLowerCase() != GimelConstants.PCATALOG_STRING && dataSet.split('.').length == 2
-  }
-
-  /**
-   * For a given dataset (table) this function calls getDataSetProperties which calls catalog provider and returns the dataset properties
-   * which will be used to identify the storage of the dataset
-   *
-   * @param datasetName  - Incoming dataset
-   * @param sparkSession - spark session
-   * @param options      - Set of user options
-   * @return - Returns the storage system (hive/teradata/kafka...)
-   */
-  def getSystemType(datasetName: String, sparkSession: SparkSession,
-                    options: Map[String, String]): DataSetType.Value = {
-    logger.info("Data set name is  ==> " + datasetName)
-    val formattedProps: Map[String, Any] = com.paypal.gimel.common.utilities.DataSetUtils.getProps(options) ++
-      Map(CatalogProviderConfigs.CATALOG_PROVIDER ->
-        sparkSession.conf.get(CatalogProviderConfigs.CATALOG_PROVIDER,
-          CatalogProviderConstants.PRIMARY_CATALOG_PROVIDER))
-
-    // if storage type unknown we will default to HIVE PROVIDER
-    if (isStorageTypeUnknown(datasetName)) {
-      formattedProps ++ Map(CatalogProviderConfigs.CATALOG_PROVIDER -> CatalogProviderConstants.HIVE_PROVIDER)
-    }
-
-    val dataSetProperties: DataSetProperties = CatalogProvider.getDataSetProperties(datasetName, options)
-    logger.info("dataSetProperties  ==> " + dataSetProperties.toString())
-    val systemType = getSystemType(dataSetProperties)
-    systemType
-  }
 }
 
 /**
@@ -1144,7 +1039,3 @@ private class DataSetOperationException(message: String, cause: Throwable)
   def this(message: String) = this(message, null)
 }
 
-object DataSetType extends Enumeration {
-  type SystemType = Value
-  val KAFKA, HBASE, HDFS, ES, HIVE, JDBC, SHERLOCK, CASSANDRA, AEROSPIKE, DRUID, RESTAPI, PIT, MONGODB, SFTP, KAFKA2, S3 = Value
-}
